unit tests: Create/provide function running-all-tests?
A public wrapper function for *running-all-tests?* allows us to create
buttons that know whether the user is running all of the unit tests
and offer to either stop or resume the unit tests, as appropriate.

I've attached a patch to the bug listed below that implements this in
the iml libs.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?11448