(Ubuntu-specific) Various hacks to get engine compiling
This commit contains various hacks to get the Halyard compiling on
Ubuntu. This commit should not be merged upstream as almost certainly
breaks compilation on other platforms and at the very least contains a
lot of "fix-this" comments which should be stripped-out.

Any suggestions on how to rewrite this commit into upstream-mergeable
commits would be appreciated.