Icons: Use icons-to-xpm.sh to convert ICO/BMP icons -> XPM
This commit uses the new icons-to-xpm.sh shell script to properly
convert the existing source icons to XPM files.

Several of the XPM files have had significant changes. This is largely
due to the fact that ImageMagic appears to use a different encoding
than whatever we used to create the XPMs that currently exist. Even
though there are several lines of changes in the XPM diffs, the data
and the display of the data should remain unchanged.

This commit also creates a couple of new XPM files as there were a
couple of ICO files in wx/src that had never been converted to XPM. I
don't believe that these new files will create any problems for
anyone.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17047